Change int to long Add help freaking everywhere Add default balance property Add balance-format placeholder Restring transfering money to yourself

+++ api/src/main/java/org/bigcraft/infpoints/api/PointType.java
@@ -4,10 +4,10 @@ import java.util.UUID;
 
 public interface PointType {
 
-    int get(UUID player);
-    void add(UUID player, int amount);
-    boolean subtract(UUID player, int amount);
-    void set(UUID player, int amount);
-    boolean transfer(UUID sender, UUID receiver, int amount);
+    long get(UUID player);
+    void add(UUID player, long amount);
+    boolean subtract(UUID player, long amount);
+    void set(UUID player, long amount);
+    boolean transfer(UUID sender, UUID receiver, long amount);
 
 }
